# api/internal/server/assets/openapi.yaml
openapi: 3.1.0
info:
  title: adVideo — API de Integrações
  version: "1.0.0"
  summary: Download de mídia de várias plataformas, consumido por outro sistema.
  description: |
    Esta é a API que **outros sistemas** usam para pedir downloads à plataforma.
    É a mesma engrenagem que atende a interface web — mesma fila, mesmos
    provedores, mesmo armazenamento —, com autenticação por chave de API em vez
    de login, e com cota, limites e auditoria por integração.

    ## Como funciona, em uma frase

    Você faz `POST /v1/integration/downloads` com uma URL, recebe `202` com um
    identificador, e nós avisamos o seu sistema por **webhook** quando o arquivo
    fica pronto. Enquanto isso, `GET /v1/integration/downloads/{id}` responde o
    estado atual a qualquer momento.

    ## Autenticação

    Toda chamada exige a chave de API emitida no painel:

    ```
    Authorization: Bearer adk_live_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
    ```

    Se o seu ambiente reescreve o header `Authorization` (proxies corporativos
    às vezes fazem isso), use `X-API-Key` com o mesmo valor.

    Opcionalmente, envie também `X-Integration-Id` com o identificador público
    da integração. Ele não substitui a chave: serve como trava contra o acidente
    de apontar a chave de homologação para a integração de produção — se os dois
    não corresponderem, a chamada é recusada com `integration_mismatch`.

    **A chave aparece uma única vez**, na resposta que a criou. O banco guarda
...

...

...

...

...
servers:
...
security:
  - ChaveDeAPI: []
...
paths:
...
  /v1/integration/downloads/{id}/cancel:
    parameters:
      - { $ref: "#/components/parameters/DownloadID" }
    post:
      tags: [Downloads]
      summary: Cancelar um download
      description: |
        Interrompe um download em andamento. O processo que está baixando é
        avisado e para em poucos segundos.

        **É idempotente.** Cancelar algo que acabou de terminar sozinho não é
        erro: a resposta devolve o status que de fato vigora, e pode ser
        `COMPLETED`. Verifique o `status` da resposta em vez de assumir
        `CANCELED`.
      operationId: cancelarDownload
      responses:
        "200":
          description: Pedido registrado; o download traz o status vigente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  download: { $ref: "#/components/schemas/Download" }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

  # -------------------------------------------------------------------------
...
components:
  securitySchemes:
    ChaveDeAPI:
      type: http
      scheme: bearer
      description: |
        Chave de API emitida no painel, com o prefixo `adk_live_`.

        Alternativa para ambientes que reescrevem `Authorization`: o header
        `X-API-Key` com o mesmo valor.
  parameters:
    DownloadID:
      name: id
      in: path
      required: true
      description: Identificador do download.
      schema: { type: string, format: uuid }
...
  responses:
    NaoAutenticado:
      description: |
        Credencial ausente, inválida, revogada ou expirada. Códigos:
        `missing_credentials`, `invalid_api_key`, `key_revoked`, `key_expired`,
        `integration_mismatch`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "chave de API inválida", code: invalid_api_key }
...
    NaoEncontrado:
      description: |
        O recurso não existe **ou não pertence a esta integração** — as duas
        situações respondem igual, de propósito: distinguir permitiria descobrir
        os identificadores de outra integração por tentativa.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "download não encontrado", code: not_found }
    LimiteExcedido:
      description: Um dos limites da integração foi atingido.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Segundos a aguardar.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: "limite de chamadas por minuto excedido; aguarde 12s"
            code: rate_limited
  schemas:
    ErrorCode:
      type: string
      description: |
        Código estável do erro. **Trate isto, não a mensagem.**
      enum:
        - missing_credentials
        - invalid_api_key
        - key_revoked
        - key_expired
        - integration_mismatch
        - integration_disabled
        - ip_not_allowed
        - rate_limited
        - quota_exceeded
        - too_many_concurrent
        - file_too_large
        - invalid_request
        - not_found
        - method_not_allowed
        - download_not_ready
        - download_expired
        - internal_error
        - service_unavailable
        - invalid_url
        - unsupported_platform
        - content_unavailable
        - content_private
        - auth_required
        - geo_blocked
        - live_content
        - format_unavailable
        - unavailable
        - timeout
        - download_failed
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Mensagem em português, para o humano que lê o log.
        code:
          $ref: "#/components/schemas/ErrorCode"
        limit:
          type: integer
          description: Presente em `file_too_large`, em bytes.
        size:
          type: integer
          description: Presente em `file_too_large`, em bytes.
    DownloadStatus:
      type: string
      description: |
        - `PENDING` — na fila
        - `PROCESSING` — baixando
        - `COMPLETED` — pronto para buscar
        - `FAILED` — falhou; veja `error_message`
        - `CANCELED` — cancelado a seu pedido
        - `EXPIRED` — a validade passou e o arquivo foi removido
        - `RETRYING` — nova tentativa após falha passageira
      enum: [PENDING, PROCESSING, COMPLETED, FAILED, CANCELED, EXPIRED, RETRYING]
...
    Progress:
      type: object
      description: Andamento. Presente enquanto o download está em curso.
      properties:
        percent: { type: number, format: double, examples: [42.5] }
        downloaded_bytes: { type: integer, format: int64 }
        total_bytes:
          type: integer
          format: int64
          description: Estimativa, corrigida conforme o download avança.
        speed_bps: { type: integer, format: int64, description: Bytes por segundo. }
        eta_seconds: { type: integer }
        postprocess:
          type: boolean
          description: |
            Verdadeiro quando os bytes já foram baixados e o arquivo está sendo
            convertido ou as faixas estão sendo juntadas. O percentual pode
            ficar em 100 nesta etapa.
    DownloadLinks:
      type: object
      description: |
        Rotas desta API para o download. As de arquivo só aparecem quando há
        arquivo — em um download que falhou, elas estão ausentes.
      properties:
        self: { type: string, format: uri }
        download_url: { type: string, format: uri }
        file: { type: string, format: uri }
    Download:
      type: object
      description: |
        O download. **Este é exatamente o mesmo objeto que vai no corpo do
        webhook**, em `data.download`: um parser serve para os dois.
      required: [id, status, title, original_url, platform, format]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/DownloadStatus" }
        title: { type: string, description: Título vindo da plataforma. }
        original_url: { type: string, format: uri }
        platform:
          type: string
          description: Identificador da plataforma.
          examples: ["youtube"]
        platform_label:
          type: string
          description: Nome da plataforma para exibição.
          examples: ["YouTube"]
        format:
          type: string
          enum: [MP4, MP3, M4A, WEBM, FLAC, BEST]
        quality_label:
          type: string
          description: Qualidade entregue, como mostrada ao usuário.
          examples: ["1080p"]
        uploader: { type: string, description: Autor ou canal. }
        thumbnail_url: { type: string, format: uri }
        duration_seconds: { type: integer }
        file_size_bytes:
          type: integer
          format: int64
          description: Tamanho real do arquivo. Só após a conclusão.
        error_message:
          type: string
          description: |
            Motivo da falha, em linguagem de usuário. O detalhe técnico não é
            exposto: ele descreve a nossa infraestrutura.
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        expires_at:
          type: string
          format: date-time
          description: Até quando o arquivo fica disponível.
        progress: { $ref: "#/components/schemas/Progress" }
        links: { $ref: "#/components/schemas/DownloadLinks" }
...